requirements.txt
args==0.1.0
click==6.6
clint==0.5.1
dotmap==1.2.0
Flask==0.11.1
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.1
MarkupSafe==0.23
pkginfo==1.3.2
pymongo==3.3.1
PyYAML==3.11
requests==2.10.0
requests-toolbelt==0.7.0
twine==1.7.4
Werkzeug==0.11.10
selenium>=2.50.0
uwsgi==2.0.14
behave